+++ Lombiq.Tests.UI.AppExtensions/SqlQueryMonitoring/ISqlQueryMonitoringStore.cs
@@ -1,0 +1,22 @@
+﻿namespace Lombiq.Tests.UI.AppExtensions.SqlQueryMonitoring;
+
+/// <summary>
+/// Stores SQL query monitoring summaries for the current tenant scope.
+/// </summary>
+public interface ISqlQueryMonitoringStore
+{
+    /// <summary>
+    /// Adds a completed monitoring summary to the store.
+    /// </summary>
+    void AddSummary(SqlQueryMonitoringSummary summary);
+
+    /// <summary>
+    /// Removes and returns the most recent summary from the store, if any.
+    /// </summary>
+    bool TryDequeueLatest(out SqlQueryMonitoringSummary summary);
+
+    /// <summary>
+    /// Clears all stored summaries.
+    /// </summary>
+    void Clear();
+}
